In Game Changes
New Script: sprite_control.cs will control the food sprite that the
turtles will eat.

Food Sprite: Food sprite will appear in a puff of smoke and drop to the
top of the water. Then the food will float for 3 seconds and then sink.
After sinking for X seconds, the food will expire and dirty the tank.